:fire: outdated warning (#1443)
Internally we have a kube setup that scales well for this now without app changes. We should at some point update this README with some basic pointers for how to deploy in abstract and maybe a link to our internal docs, but for now we can just kill this warning.