Update currencies
The list of currencies in `Shopify/quilt` has not been kept in sync with the list in `Shopify/shopify`.  After reviewing the two lists, I've identified a handful of codes that need to be included so that we have the same version of "the truth" in both places.

New currencies:
  - `BYN` Belarusia New Ruble
  - `BMD` Bermuda Dollar (pegged 1:1 to USD)
  - `DJF` Djibouti Franc
  - `ERN` Eritrea Nakfa
  - `FKP` Falkland Islands Pound (pegged 1:1 to GBP)
  - `GIP` Gibraltar Pound (pegged 1:1 to GBP)
  - `GNF' Guinea Franc
  - `IRR` Iran Rial
  - `KID` Kiribati Dollar (pegged 1:1 to AUD)
  - `LYD` Libya Dinar
  - `MRU` Mauritania New (Second) Ouguiya
  - `PAB` Panama Balboa
  - `SHP` Saint Helena Pound (pegged 1:1 to GBP)
  - `SLL` Sierra Leone Leone
  - `SOS` Somalia Shilling (a fiat currency that is unique because it's no longer backed by a central bank)
  - `TJS` Tajikistan Simoni
  - `TOP` Tonga Pa'anga
  - `VES` Venezuelan Sovereign Bolivar

Historical currencies that are being moved to the end of the list:
  - `BYR` Belarusia (Old) Ruble (replaced by `BYN`)
  - `VEF` Venezuealan Strong Bolivar (replaced by `VES`)