Always output a valid es module file (#131)
In the unlikely event that a schema is borderline empty and wouldn't
produce any types we previously outputted a blank file. Which causes
type checking with isolatedModules enabled to fail because a blank file
is ambigiuous - it coould be a script or a module. This change ensures
that if a file would otherwise be blank we output a `export {}` to make
it clear that the file should be treated as an es modules file.