Fix setting state in IntersectionObserver render
When a consumer of `@shopify/react-intersection-observer` attempts to set state in the `onIntersectionChange` callback for `<IntersectionObserver>`, React 16.13 and above will generate a warning that the consuming component cannot update state while `IntersectionObserver` is rendering. This is because the `useValueTracking` hook used by `IntersectionObserver` is updating it's tracking ref and triggering the callback during the render.

The fix is to simply wrap that logic in a `useEffect` hook so that it happens after the render.

Fixes https://github.com/Shopify/quilt/issues/2058