Update react-testing to use new createRoot API
The previous ReactDOM.render method is deprecated and instead tests
should be making use of createRoot to generate a "Root" element and then
calling render/unmount on that.

This causes some confusion here because we already have a concept of a
"Root" in react-testing, and so we use the name "reactRoot" to
differentiate here.