Fix subtle test issues in react-form and react-form-state
These two test suites weren't always awaiting the promises that got
enqueued as part of the actions they were triggering. This can lead to
bugs when upgrading to React 18 because of the way that renders are
changed and we need to use the correct process in order to ensure that
we're ending up testing the right state of the component.